Rename top-level package to make stubs discoverable
Rename the top-level Python package to "flask_httpauth-stubs" so that
Mypy (and other possibly other type checkers) can find the stubs.

This is the correct name, since the top-level package name for the code
installed from Flask-HTTPAuth distribution package is "flask_httpauth"
and the top-level package installed from the corresponding stubs
distribution package must be exactly the same with "-stubs" appended.

Also name the distribution package to "Flask-HTTPAuth-stubs" so that it
is in line with the naming of the "Flask-HTTPAuth" distribution package.

Note: The term Distribution Package must not be confused with the term
Python Package (or Import Package).  See e.g. [1] for more details.

[1] https://packaging.python.org/glossary/#term-Distribution-Package